Resend Email for User and Provider, Send Email verification on signup, Facebook Login verification, Fare Calculation

+++ database/migrations/2017_01_11_182738_create_service_types_table.php
@@ -18,8 +18,8 @@ class CreateServiceTypesTable extends Migration
             $table->string('name');
             $table->string('provider_name')->nullable();
             $table->string('image')->nullable();
-            $table->integer('fixed')->nullable();
-            $table->integer('price')->nullable();
+            $table->double('fixed')->nullable();
+            $table->double('price')->nullable();
             $table->string('description')->nullable();
             $table->integer('status')->default(0);
             $table->timestamps();
